requirements.txt
# pip install -r requirements.txt
matplotlib>=3.2.2
numpy>=1.18.5
opencv-python>=4.1.2
torch>=1.6.0
torchvision>=0.7.0
tqdm>=4.41.0
pillow